🎨 Move ChangeElection to its own component file
