+++ src/components/list/UserBar/userBar.css
@@ -36,67 +36,4 @@
       align-items: center;
     }
   }
-
-  .change-election {
-    height: max-content;
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    margin: auto;
-    padding: 20px;
-    border-radius: 10px;
-    background-color: black;
-    color: white;
-    border: none;
-    cursor: pointer;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    img {
-      width: 35px;
-      height: 35px;
-      padding-right: 10px;
-    }
-
-    .election-list {
-      flex: 1;
-      display: flex;
-      flex-direction: column;
-      min-height: 200px;
-      gap: 20px;
-      width: 100%;
-
-      ul {
-        padding: 20px;
-        border-radius: 10px;
-        background-color: rgba(17, 25, 40, 0.6);
-        color: white;
-        border: none;
-        cursor: pointer;
-        list-style-type: none;
-        gap: 20px;
-        width: 100%;
-
-        li {
-          width: 100%;
-          padding-top: 20px;
-          display: flex;
-          align-items: center;
-          gap: 20px;
-        }
-
-        button {
-          padding: 10px;
-          border-radius: 10px;
-          background-color: #1a73e8;
-          color: white;
-          border: none;
-          cursor: pointer;
-        }
-      }
-    }
-  }
 }
